+++ sources/frigid/__/doctab.py
@@ -20,8 +20,6 @@
 
 ''' Docstrings table for reuse across entities. '''
 
-# ruff: noqa: F403,F405
-
 
 from . import imports as __